Basic data fetch capability + major refactoring.
